+++ src/test/java/seedu/restaurant/commons/events/ui/menu/ItemPanelSelectionChangedEventTest.java
@@ -1,4 +1,4 @@
-package seedu.restaurant.commons.events.ui;
+package seedu.restaurant.commons.events.ui.menu;
 
 import static org.junit.Assert.assertEquals;
