+++ src/test/java/seedu/restaurant/commons/events/ui/menu/JumpToItemListRequestEventTest.java
@@ -1,4 +1,4 @@
-package seedu.restaurant.commons.events.ui;
+package seedu.restaurant.commons.events.ui.menu;
 
 import static org.junit.Assert.assertEquals;
 
@@ -6,7 +6,6 @@ import org.junit.Test;
 
 import seedu.restaurant.commons.core.index.Index;
 import seedu.restaurant.commons.events.BaseEvent;
-import seedu.restaurant.commons.events.ui.menu.JumpToItemListRequestEvent;
 
 public class JumpToItemListRequestEventTest {
