+++ src/test/java/seedu/restaurant/commons/events/ui/reservation/DisplayReservationListRequestEventTest.java
@@ -1,4 +1,4 @@
-package seedu.restaurant.commons.events.ui;
+package seedu.restaurant.commons.events.ui.reservation;
 
 import static org.junit.Assert.assertEquals;
